Split expansion dependency tooltip component
